Started working on a Vulkan renderer API, still has several issues.

+++ Waffle/src/Waffle/Renderer/Framebuffer.h
@@ -59,7 +59,7 @@ namespace Waffle {
 
 		virtual void ClearAttachment(uint32_t attachmentIndex, int value) = 0;
 
-		virtual uint32_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
+		virtual uint64_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
 
 		virtual const FramebufferSpecification& GetSpecification() const = 0;
